refactor: extract delegated skill evidence
