+++ crates/agent-engine/src/runtime/virtual_tools_tests.rs
@@ -5,12 +5,13 @@ use crate::{
     rollout::{RolloutItem, RolloutRecorder},
     runtime::{
         ChildRunRecord, ChildRunStatus, NamespaceRuntimeEnvironment, RuntimeConfig, TurnState,
+        agent_loop::NamespaceActionRecord, delegated_child_run::MAX_DELEGATED_RESULT_SUMMARY_CHARS,
         turn_state::TurnActivityState,
     },
     skills::{
-        ActiveSkillEnvelope, ResolvedCapabilityView, ResolvedSkillExecution, ScopedPackageDir,
-        SkillActivationReason, SkillExecutionResolutionSource, SkillHostCapabilities,
-        SkillMetadata, SkillScope,
+        ActiveSkillEnvelope, DelegatedSkillInvocationRecord, ResolvedCapabilityView,
+        ResolvedSkillExecution, ScopedPackageDir, SkillActivationReason,
+        SkillExecutionResolutionSource, SkillHostCapabilities, SkillMetadata, SkillScope,
     },
     tools::ToolRegistry,
 };
@@ -285,31 +286,6 @@ fn test_update_plan_tool_definition() {
     assert_eq!(def.parameters["type"], "object");
     assert!(def.parameters["properties"].get("explanation").is_some());
     assert!(def.parameters["properties"].get("items").is_some());
-}
-
-#[test]
-fn test_invoke_delegated_skill_tool_definition() {
-    let def = invoke_delegated_skill_tool_definition();
-    assert_eq!(def.name, "invoke_delegated_skill");
-    assert!(def.description.contains("delegated skill"));
-    assert_eq!(def.parameters["type"], "object");
-    assert_eq!(def.parameters["properties"]["skill_id"]["type"], "string");
-    assert_eq!(
-        def.parameters["properties"]["skill_id"]["maxLength"],
-        MAX_DELEGATED_SKILL_ID_CHARS
-    );
-    assert_eq!(def.parameters["properties"]["target"]["type"], "string");
-    assert_eq!(
-        def.parameters["properties"]["target"]["maxLength"],
-        MAX_DELEGATED_TARGET_CHARS
-    );
-    assert_eq!(def.parameters["properties"]["task"]["type"], "string");
-    assert_eq!(
-        def.parameters["properties"]["task"]["maxLength"],
-        MAX_DELEGATED_TASK_CHARS
-    );
-    assert!(def.parameters["properties"].get("workspace_root").is_none());
-    assert_eq!(def.parameters["properties"]["cwd"]["type"], "string");
 }
 
 #[test]
@@ -925,238 +901,6 @@ fn test_parse_plan_update_using_description() {
     assert_eq!(result.unwrap().1[0].content, "Step description");
 }
 
-#[test]
-fn test_parse_delegated_skill_invocation_request_valid() {
-    let args = json!({
-        "skill_id": "repo-review",
-        "target": "reviewer",
-        "task": "Review the current diff and summarize risks."
-    });
-
-    let result = parse_delegated_skill_invocation_request(&args).unwrap();
-    assert_eq!(result.skill_id, "repo-review");
-    assert_eq!(result.target, "reviewer");
-    assert_eq!(result.task, "Review the current diff and summarize risks.");
-}
-
-#[test]
-fn test_parse_delegated_skill_invocation_request_treats_empty_optional_cwd_as_absent() {
-    let args = json!({
-        "skill_id": "repo-review",
-        "target": "reviewer",
-        "task": "Review the current diff and summarize risks.",
-        "cwd": "   "
-    });
-
-    let result = parse_delegated_skill_invocation_request(&args).unwrap();
-    assert_eq!(result.cwd, None);
-}
-
-#[test]
-fn test_parse_delegated_skill_invocation_request_rejects_empty_fields() {
-    let missing = json!({
-        "skill_id": "repo-review",
-        "target": "reviewer"
-    });
-    assert!(parse_delegated_skill_invocation_request(&missing).is_none());
-
-    let empty = json!({
-        "skill_id": "repo-review",
-        "target": "reviewer",
-        "task": "   "
-    });
-    assert!(parse_delegated_skill_invocation_request(&empty).is_none());
-}
-
-#[test]
-fn test_parse_delegated_skill_invocation_request_accepts_bounded_timeout() {
-    let request = parse_delegated_skill_invocation_request(&json!({
-        "skill_id": "repo-review",
-        "target": "reviewer",
-        "task": "Review the current diff.",
-        "timeout_secs": 600
-    }))
-    .expect("expected delegated request");
-
-    assert_eq!(request.timeout_secs, Some(600));
-}
-
-#[test]
-fn test_parse_delegated_skill_invocation_request_rejects_invalid_timeout() {
-    assert!(
-        parse_delegated_skill_invocation_request(&json!({
-            "skill_id": "repo-review",
-            "target": "reviewer",
-            "task": "Review the current diff.",
-            "timeout_secs": 0
-        }))
-        .is_none()
-    );
-    assert!(
-        parse_delegated_skill_invocation_request(&json!({
-            "skill_id": "repo-review",
-            "target": "reviewer",
-            "task": "Review the current diff.",
-            "timeout_secs": (MAX_DELEGATED_TIMEOUT_SECS + 1)
-        }))
-        .is_none()
-    );
-}
-
-#[test]
-fn test_build_bounded_delegated_invocation_persistence_truncates_fields() {
-    let request = DelegatedSkillInvocationRequest {
-        skill_id: "s".repeat(MAX_DELEGATED_SKILL_ID_CHARS + 40),
-        target: "t".repeat(MAX_DELEGATED_TARGET_CHARS + 40),
-        task: "x".repeat(MAX_DELEGATED_TASK_CHARS + 200),
-        cwd: Some(PathBuf::from(format!(
-            "/tmp/{}",
-            "c".repeat(MAX_DELEGATED_PATH_CHARS + 20)
-        ))),
-        timeout_secs: Some(DEFAULT_DELEGATED_TIMEOUT_SECS),
-    };
-    let result = DelegatedSkillResult::failed(
-        format!(
-            "Delegated skill '{}' resolved to delegated target '{}', but delegated launch support is not yet available in this runtime.",
-            request.skill_id, request.target
-        ),
-        Some(json!({
-            "error_kind": "runtime_child_launch_unavailable"
-        })),
-    );
-
-    let child_run = Some(DelegatedChildRunReference {
-        process_path: "/proc/42".to_string(),
-        child_run_id: None,
-        state_ref: None,
-        rollout_debug_path: None,
-        terminal_status: "completed".to_string(),
-    });
-    let (arguments, record, rollout_record) =
-        build_bounded_delegated_invocation_persistence(&request, result, child_run);
-
-    let skill_id = arguments["skill_id"].as_str().unwrap();
-    let target = arguments["target"].as_str().unwrap();
-    let task = arguments["task"].as_str().unwrap();
-    assert!(skill_id.chars().count() <= MAX_DELEGATED_SKILL_ID_CHARS);
-    assert!(target.chars().count() <= MAX_DELEGATED_TARGET_CHARS);
-    assert!(task.chars().count() <= MAX_DELEGATED_TASK_CHARS);
-    assert!(skill_id.ends_with("..."));
-    assert!(target.ends_with("..."));
-    assert!(task.ends_with("..."));
-    assert!(arguments.get("workspace_root").is_none());
-    assert!(arguments["cwd"].as_str().unwrap().chars().count() <= MAX_DELEGATED_PATH_CHARS);
-    assert_eq!(
-        arguments["timeout_secs"].as_u64(),
-        Some(DEFAULT_DELEGATED_TIMEOUT_SECS)
-    );
-    assert!(record.result.summary.chars().count() <= MAX_DELEGATED_RESULT_SUMMARY_CHARS);
-    assert!(record.result.summary.ends_with("..."));
-    assert_eq!(
-        rollout_record.child_run.as_ref().unwrap().process_path,
-        "/proc/42"
-    );
-}
-
-#[test]
-fn test_build_bounded_delegated_invocation_persistence_keeps_child_run_out_of_tape_record() {
-    let request = DelegatedSkillInvocationRequest {
-        skill_id: "repo-review".to_string(),
-        target: "reviewer".to_string(),
-        task: "Review the current diff and summarize risks.".to_string(),
-        cwd: Some(PathBuf::from("/mnt/source/src")),
-        timeout_secs: Some(600),
-    };
-    let result = DelegatedSkillResult::completed("Delegated review completed.", None);
-    let child_run = Some(DelegatedChildRunReference {
-        process_path: "/proc/42".to_string(),
-        child_run_id: None,
-        state_ref: None,
-        rollout_debug_path: Some("/tmp/inline-child.jsonl".to_string()),
-        terminal_status: "completed".to_string(),
-    });
-
-    let (_, tape_record, rollout_record) =
-        build_bounded_delegated_invocation_persistence(&request, result, child_run);
-    let tape_payload = serde_json::to_value(&tape_record).unwrap();
-    let rollout_payload = serde_json::to_value(&rollout_record).unwrap();
-
-    assert!(tape_payload.get("child_run").is_none());
-    assert_eq!(
-        rollout_payload["child_run"]["process_path"],
-        json!("/proc/42")
-    );
-    assert_eq!(
-        rollout_payload["child_run"]["rollout_debug_path"],
-        json!("/tmp/inline-child.jsonl")
-    );
-    assert!(rollout_payload["child_run"].get("rollout_path").is_none());
-    assert!(tape_payload.get("workspace_root").is_none());
-    assert_eq!(tape_payload["cwd"], json!("/mnt/source/src"));
-    assert_eq!(tape_payload["timeout_secs"], json!(600));
-}
-
-#[test]
-fn test_build_bounded_delegated_invocation_persistence_bounds_result_sidecars() {
-    let request = DelegatedSkillInvocationRequest {
-        skill_id: "repo-review".to_string(),
-        target: "reviewer".to_string(),
-        task: "Review the current diff and summarize risks.".to_string(),
-        cwd: Some(PathBuf::from("/mnt/source/src")),
-        timeout_secs: None,
-    };
-    let mut result = DelegatedSkillResult::failed("Delegated review failed.", None);
-    result.child_run = Some(json!({
-        "id": "child-run-1",
-        "status": "failed",
-        "warnings": vec!["child-warning".repeat(200); MAX_DELEGATED_RESULT_WARNINGS + 8],
-        "large_metadata": "x".repeat(MAX_DELEGATED_CHILD_RUN_METADATA_CHARS * 2)
-    }));
-    result.warnings = (0..(MAX_DELEGATED_RESULT_WARNINGS + 3))
-        .map(|index| {
-            format!(
-                "warning-{index:03}-{}",
-                "x".repeat(MAX_DELEGATED_RESULT_WARNING_CHARS)
-            )
-        })
-        .collect();
-
-    let (_, tape_record, _) =
-        build_bounded_delegated_invocation_persistence(&request, result, None);
-
-    assert_eq!(
-        tape_record.result.warnings.len(),
-        MAX_DELEGATED_RESULT_WARNINGS
-    );
-    assert!(tape_record.result.warnings[0].starts_with("warning-003-"));
-    assert!(
-        tape_record
-            .result
-            .warnings
-            .iter()
-            .all(|warning| warning.chars().count() <= MAX_DELEGATED_RESULT_WARNING_CHARS)
-    );
-    assert!(tape_record.result.warnings.last().unwrap().ends_with("..."));
-    assert!(
-        tape_record
-            .result
-            .child_run
-            .as_ref()
-            .unwrap()
-            .to_string()
-            .len()
-            <= MAX_DELEGATED_CHILD_RUN_METADATA_CHARS
-    );
-    let truncation = tape_record.result.truncation.unwrap();
-    assert!(truncation.child_run);
-    assert!(truncation.warnings);
-    assert!(truncation.original_child_run_chars.unwrap() > MAX_DELEGATED_CHILD_RUN_METADATA_CHARS);
-    assert_eq!(
-        truncation.original_warning_count,
-        Some(MAX_DELEGATED_RESULT_WARNINGS + 3)
-    );
-}
-
 #[tokio::test]
 async fn redaction_expansion_preserves_delegated_output_reference_and_child_paths() {
     let (state, shell) = create_namespace_agent_loop_state_and_shell();
@@ -1207,93 +951,6 @@ async fn redaction_expansion_preserves_delegated_output_reference_and_child_path
     let retained = String::from_utf8(shell.cat(&output_ref.path).await.unwrap()).unwrap();
     assert!(retained.chars().count() > MAX_DELEGATED_RESULT_OUTPUT_INLINE_CHARS);
     assert!(retained.contains("[REDACTED reason=credential_token]"));
-}
-
-#[test]
-fn test_build_bounded_delegated_invocation_persistence_truncates_structured_output() {
-    let request = DelegatedSkillInvocationRequest {
-        skill_id: "repo-review".to_string(),
-        target: "reviewer".to_string(),
-        task: "Review the current diff and summarize risks.".to_string(),
-        cwd: Some(PathBuf::from("/tmp/repo/src")),
-        timeout_secs: None,
-    };
-    let result = DelegatedSkillResult::completed(
-        "Delegated review completed.",
-        Some(json!({
-            "status": "completed",
-            "summary": "Delegated review completed.",
-            "details": "x".repeat(MAX_DELEGATED_STRUCTURED_OUTPUT_CHARS * 2)
-        })),
-    );
-
-    let (_, tape_record, _) =
-        build_bounded_delegated_invocation_persistence(&request, result, None);
-    let structured = tape_record.result.structured_output.unwrap();
-    assert!(structured.to_string().len() <= MAX_DELEGATED_STRUCTURED_OUTPUT_CHARS);
-    assert_eq!(structured["status"], json!("completed"));
-}
-
-#[test]
-fn test_delegated_rollout_record_flattens_invocation_result() {
-    let request = DelegatedSkillInvocationRequest {
-        skill_id: "repo-review".to_string(),
-        target: "reviewer".to_string(),
-        task: "Review the current diff.".to_string(),
-        cwd: None,
-        timeout_secs: None,
-    };
-    let mut result = DelegatedSkillResult::completed("Review completed.", None);
-    result.output_ref = Some(DelegatedSkillOutputRef {
-        path: "/agent/1/actions/a0/output".to_string(),
-        offset: Some(0),
-        length: Some(42),
-        debug: Some(DelegatedSkillOutputDebugMetadata {
-            process_path: "child-machine".to_string(),
-            rollout_path: Some("/tmp/child.jsonl".to_string()),
-            field: "output_text".to_string(),
-        }),
-    });
-
-    let (_, _, rollout_record) =
-        build_bounded_delegated_invocation_persistence(&request, result, None);
-    let serialized = serde_json::to_value(rollout_record).unwrap();
-
-    assert_eq!(
-        serialized.pointer("/result/output_ref/debug/rollout_path"),
-        Some(&json!("/tmp/child.jsonl"))
-    );
-    assert!(serialized.get("invocation").is_none());
-}
-
-#[test]
-fn test_build_bounded_delegated_invocation_persistence_truncates_oversized_summary() {
-    let request = DelegatedSkillInvocationRequest {
-        skill_id: "repo-review".to_string(),
-        target: "reviewer".to_string(),
-        task: "Review the current diff and summarize risks.".to_string(),
-        cwd: Some(PathBuf::from("/tmp/repo/src")),
-        timeout_secs: None,
-    };
-    let result = DelegatedSkillResult::completed(
-        "Delegated review completed.",
-        Some(json!({
-            "status": "completed",
-            "summary": "y".repeat(MAX_DELEGATED_STRUCTURED_OUTPUT_CHARS * 2),
-            "details": "x".repeat(MAX_DELEGATED_STRUCTURED_OUTPUT_CHARS * 2)
-        })),
-    );
-
-    let (_, tape_record, _) =
-        build_bounded_delegated_invocation_persistence(&request, result, None);
-    let structured = tape_record.result.structured_output.unwrap();
-    let summary = structured["summary"]
-        .as_str()
-        .expect("summary should remain string");
-    assert!(structured.to_string().len() <= MAX_DELEGATED_STRUCTURED_OUTPUT_CHARS);
-    assert!(summary.len() < MAX_DELEGATED_STRUCTURED_OUTPUT_CHARS);
-    assert!(summary.ends_with("..."));
-    assert_eq!(structured["status"], json!("completed"));
 }
 
 // Tests for parse_plan_status
